context(primitives): flatten messages submodules into messages.rs and update imports
