--- crates/context/primitives/src/messages/sync.rs
@@ -1,13 +1,0 @@
-use actix::Message;
-use calimero_primitives::application::ApplicationId;
-use calimero_primitives::context::ContextId;
-
-#[derive(Copy, Clone, Debug)]
-pub struct SyncRequest {
-    pub context_id: ContextId,
-    pub application_id: ApplicationId,
-}
-
-impl Message for SyncRequest {
-    type Result = ();
-}
